requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
libgravatar==1.0.0
MarkupSafe==1.1.1
PyMySQL==1.0.2
Werkzeug==1.0.1
Python-dotenv=0.17.1
PyJWT==2.1.0
numpy==1.20.2
pdfkit==0.6.1
twilio==6.61.0
requests==2.26.0
pandas==1.2.1